Another hacky fix papering over a bug in SockJS.
Faye::Websocket response objects, don't have a finish method, so
currently if we reach the close method inside a Consumer from a
websocket connection, the server will force exit.

Figuring out why there is this discrepency would require some heavy duty
archaeology into the codebase, but I'm not sure it's worth it just now.
This fix just prevents us trying to call finish on an object that
doesn't respond to it.